requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
xgboost==2.1.1
requests==2.32.3
